+++ src/Oro/Bundle/WebCatalogBundle/Provider/ContentVariantProvider.php
@@ -5,11 +5,12 @@ namespace Oro\Bundle\WebCatalogBundle\Provider;
 use Doctrine\ORM\QueryBuilder;
 use Oro\Component\WebCatalog\ContentVariantProviderInterface;
 use Oro\Component\WebCatalog\Entity\ContentNodeInterface;
+use Oro\Component\WebCatalog\SortableContentVariantProviderInterface;
 
 /**
  * Delegates the getting information about a content variant to child providers.
  */
-class ContentVariantProvider implements ContentVariantProviderInterface
+class ContentVariantProvider implements SortableContentVariantProviderInterface
 {
     /** @var iterable|ContentVariantProviderInterface[] */
     private $providers;
@@ -89,6 +90,24 @@ class ContentVariantProvider implements ContentVariantProviderInterface
             $recordId = $provider->getRecordId($item);
             if ($recordId) {
                 return $recordId;
+            }
+        }
+
+        return null;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getRecordSortOrder(array $item)
+    {
+        foreach ($this->providers as $provider) {
+            if (!$provider instanceof SortableContentVariantProviderInterface) {
+                continue;
+            }
+            $recordSortOrder = $provider->getRecordSortOrder($item);
+            if (!is_null($recordSortOrder)) {
+                return $recordSortOrder;
             }
         }